Add functionality to post a new comment to an article
