+++ public/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <title>Newsly</title>
+    <title>NC / DC</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
