+++ src/index.css
@@ -317,6 +317,19 @@ input[type=text]:focus {
     font-weight: 900;
 }
 
+.alert {
+    position: relative;
+    padding: 1rem 1rem;
+    margin-bottom: 1rem;
+    border: 1px solid transparent;
+}
+
+.alert-success {
+  color: #0f5132;
+  background-color: #d1e7dd;
+  border-color: #0f5132;
+}
+
 a {
     color: #f92c85;
     text-decoration: none;
